web/pyproject.toml
[tool.poetry]
name = "conference_catering_browser"
version = "0.1.0"
description = ""
authors = ["Brian Ibbotson <brian@quinceleaf.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
gunicorn = "^20.1.0"
django-filter = "^2.4.0"
django-pandas = "^0.6.4"
django-select2 = "^7.6.2"
django-widget-tweaks = "^1.4.8"
djangorestframework = "^3.12.2"
numpy = "^1.20.2"
openpyxl = "^3.0.7"
pandas = "^1.2.3"
Pillow = "^8.1.2"
python-dotenv = "^0.15.0"
pytz = "^2021.1"
WeasyPrint = "^52.5"
XlsxWriter = "^1.3.9"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django-debug-toolbar = "^3.2.1"
flake8 = "^3.9.1"


[build-system]
requires = ["poetry-core==1.0.0","setuptools>=50", "cryptography<=3.3.2"]
build-backend = "poetry.core.masonry.api"
